+++ plone/namedfile/scaling.py
@@ -21,6 +21,7 @@ from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import safe_encode
 from Products.Five import BrowserView
 from xml.sax.saxutils import quoteattr
+from zExceptions import BadRequest
 from zExceptions import Unauthorized
 from ZODB.blob import BlobFile
 from ZODB.POSException import ConflictError
@@ -30,7 +31,7 @@ from zope.component import queryUtility
 from zope.deprecation import deprecate
 from zope.interface import alsoProvides
 from zope.interface import implementer
-from zope.publisher.interfaces import IPublishTraverse
+from zope.publisher.interfaces.browser import IBrowserPublisher
 from zope.publisher.interfaces import NotFound
 from zope.traversing.interfaces import ITraversable
 from zope.traversing.interfaces import TraversalError
@@ -390,7 +391,7 @@ class DefaultImageScalingFactory:
         return value, format_, dimensions
 
 
-@implementer(ITraversable, IPublishTraverse)
+@implementer(ITraversable, IBrowserPublisher)
 class ImageScaling(BrowserView):
     """view used for generating (and storing) image scales"""
 
@@ -433,6 +434,10 @@ class ImageScaling(BrowserView):
             )
             return scale_view
         raise NotFound(self, name, self.request)
+
+    def browserDefault(self, request):
+        # There's nothing in the path after /@@images
+        raise BadRequest("Missing image scale path")
 
     def traverse(self, name, furtherPath):
         """used for path traversal, i.e. in zope page templates"""