+++ CHANGES.rst
@@ -8,6 +8,42 @@ Changelog
 
 .. towncrier release notes start
 
+6.1.1 (2023-06-22)
+------------------
+
+Bug fixes:
+
+
+- Return a 400 Bad Request response if the `@@images` view is published without a subpath. @davisagli (#144)
+
+
+Tests
+
+
+- Fix tests to work with various ``beautifulsoup4`` versions.
+  [maurits] (#867)
+
+
+6.1.0 (2023-05-22)
+------------------
+
+New features:
+
+
+- Move ``Zope2FileUploadStorable`` code from plone.app.z3cform to here to break a cyclic dependency.
+  [gforcada] (#3764)
+
+
+6.0.2 (2023-05-08)
+------------------
+
+Bug fixes:
+
+
+- Fix picture tag when original image is used instead of a scale.
+  [maurits] (#142)
+
+
 6.0.1 (2023-03-14)
 ------------------